#16312. Add geofilt boosting. Adjust JUnit to test geofilt queries + parseKm method.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@22709 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/common-services/src/main/java/ish/oncourse/services/search/SearchService.java
@@ -37,7 +37,7 @@ public class SearchService implements ISearchService {
     /**
      * Default maximum distance for geo-searches
      */
-    public static final int MAX_DISTANCE = 100;
+	public static final double MAX_DISTANCE = 100.0;
 
     @Inject
     private IWebSiteService webSiteService;
